requirements.txt
django-cripsy-forms